gdb::option: Add support for filename option.
This commit adds support for filename option for the gdb::option parsing
framework.  It is meant to help the user enter values that could be
parsed by built_argv.

Tab completion works as expected, i.e. similar to bash's behavior.

Considering the following tree structure:

somedir/
├── a folder
└── a file

The following completions are available:

    (gdb) maintenance test-options unknown-is-error -filename somedir/a\ <tab>
    a file    a folder/

or

    (gdb) maintenance test-options unknown-is-error -filename "somedir/a <tab>
    a file    a folder/

The complete command command also provides useful completions which are
escaped or quoted depending on what the user did input initially.

    (gdb) complete maintenance test-options unknown-is-error -filename somedir/a
    maintenance test-options unknown-is-error -filename somedir/a\ file
    maintenance test-options unknown-is-error -filename somedir/a\ folder/
    (gdb) complete maintenance test-options unknown-is-error -filename 'somedir/a
    maintenance test-options unknown-is-error -filename 'somedir/a file
    maintenance test-options unknown-is-error -filename 'somedir/a folder/

Note that since entering a path is an interactive process, the completer
will not append the closing quote after a directory name. Instead it
adds a '/' and expect the user to provide further input.

Having both of those behaviors at the same time requires that the
completer knows if it has been called by readline after the user
pressed the tab key or by some other mechanism.  In the first scenario it
should display non quoted/escaped completions so the user sees actual
filenames but complete quoted/escaped filenames.  In the second scenario
it should only provide quoted/escaped completions.  For that purpose,
this commit adds a m_from_readline property to the completion_tracker
object and initializes it accordingly to the calling context.

This filename_completer has slightly different behavior than the one
available in gdb/completer.c mainly with regard to escaping/quoting.
The original completer is kept as is to maintain stable behavior for
commands whose completer might depend on this implementation.  If this is
desirable, I’ll happily replace the existing filename completer with the
one currently proposed in the gdb::options::filename namespace instead
of having both with slightly different behaviors (but I an not sure if
users are relying on the current behavior of the filename completer).

It really feels that I have had to re-implement many existing readline
behavior and I would be glad if someone has a simpler approach to obtain
the same result.

All feedback are welcome.

gdb/ChangeLog:

	* cli/cli-option.c (union option_value): Add filename option.
	(struct option_def_and_value): Free filename in dtor.
	(gdb::option::filename::quote_filename): New function.
	(gdb::option::filename::handle_completion): New function.
	(gdb::option::filename::filename_copleter): New function.
	(gdb::option::parse_option): Add support for filename.
	(gdb::option::save_option_value_in_ctx): Add support for filename.
	(gdb::option::get_val_type_str): Add support for filename.
	(gdb::option::add_setshow_cmds_for_options): Add support for filename.
	* cli/cli-option.h (struct option_def): Add filename field.
	(struct gdb::option::filename_option_def): Create.
	* completer.c (completion_tracker::discard_completions): Reset the
	m_from_readline flag.
	(gdb_completion_word_break_characters_throw): Set the m_from_readline
	flag on the completion_tracker.
	* completer.h (class completion_tracker): Add m_from_readline fileld
	and associated getter / setters.
	* maint-test-options.c (struct test_options_opts): Add a -filename
	option.

gdb/testsuite/ChangeLog:

	* gdb.base/options.exp: Add tests for the filename option.
	* lib/completion-support.exp (make_tab_completion_list_re):
	Relax number of required spaced at the end of the pattern.